+++ src/client/local_server_interface.cpp
@@ -85,7 +85,7 @@ void LocalServerInterface::setPlayerMoveInput(int moveInput) {
 		player->setMoveInput(moveInput);
 }
 
-void LocalServerInterface::setPlayerOrientation(double yaw, double pitch) {
+void LocalServerInterface::setPlayerOrientation(float yaw, float pitch) {
 	player->setOrientation(yaw, pitch);
 }
